Complete restructure Restructure the Server trait for simpler operation with command line Restructure tests for better reusability

+++ src/main.rs
@@ -1,24 +1,22 @@
-#![allow(warnings)]
+// #![allow(warnings)]
 
-// slint::slint!(import { AnyServeUI } from "src/ui/ui.slint";);
 
-use hyper::client::connect;
-use log::{info, warn, error, debug, LevelFilter};
+use log::{error, info, warn};
 
 use std::path::PathBuf;
 use std::ops::Deref;
 use std::sync::Arc;
-use std::{env, process};
 
-use tokio::sync::broadcast;
-use futures::future::join_all;
+
+
+
 
 mod tests;
 mod utils;
 mod servers;
 use crate::servers::{*};
 
-use clap::{Args, Parser, Subcommand};
+use clap::{Parser};
 extern crate ctrlc;
 
 #[derive(Parser, Debug)]
@@ -34,8 +32,8 @@ struct Cli {
     )] bind_ip: String,
     
     #[arg(
-        help = "Directory to serve",
-        short, long, required = false,
+        help = "Path to serve",
+        short = 'p', long, required = false,
         default_value = "/tmp/",
         value_name = "DIR",
         require_equals = true,
@@ -84,18 +82,24 @@ struct Cli {
 
 #[tokio::main]
 async fn main() {
-    ::std::env::set_var("RUST_LOG", "debug");
+    let cli_args = Cli::parse();
+
+    let mut log_level = "info";
+    if cli_args.verbose > 0 {
+        log_level = "debug";
+    }
+
+    ::std::env::set_var("RUST_LOG", log_level);
     env_logger::builder()
         .format_timestamp_secs()
         .init();
 
     //////////////////////////////////////////////////////////////////////
     //////////////////////////////////////////////////////////////////////
-    let cli_args = Cli::parse();
-    println!("{:#?}", cli_args);
+    // debug!("\n{:#?}\n", cli_args);
 
+    let mut spawned_runners = vec![];
     let mut spawned_servers = vec![];
-    let mut servers = vec![];
 
     // Read and validate the bind address
     let bind_ip = cli_args.bind_ip;
@@ -107,21 +111,19 @@ async fn main() {
     // TFTP from here on
     // 
     // Spin the runners to wait for any potential server start
-    let tftp_server = Arc::new(<Server as TFTPServerRunner>::new());
-    let tftp_server_c = tftp_server.clone();
-
-    servers.push(tftp_server.clone());
-
     if cli_args.tftp.is_some() {
-        spawned_servers.push(
+        let tftp_server = Arc::new(<Server as TFTPServerRunner>::new());
+        let tftp_server_c = tftp_server.clone();
+
+        spawned_servers.push(tftp_server.clone());
+        spawned_runners.push(
             tokio::spawn(async move {
-            TFTPServerRunner::runner(tftp_server.deref()).await
+                TFTPServerRunner::runner(tftp_server).await
             })
         );
 
-        let port = cli_args.tftp.unwrap() as u16;
-
-        tftp_server_c.start(path.clone(), bind_ip.clone(), port);
+        let _port = cli_args.tftp.unwrap() as u16;
+        let _ = tftp_server_c.start();
     }
 
 
@@ -130,21 +132,24 @@ async fn main() {
     // FTP from here on
     // 
     // Spin the runners to wait for any potential server start
-    let ftp_server = Arc::new(<Server as FTPServerRunner>::new());
-    let ftp_server_c = ftp_server.clone();
-
-    servers.push(ftp_server.clone());
-
     if cli_args.ftp.is_some() {
-        spawned_servers.push(
+        let port = cli_args.ftp.unwrap() as u16;
+        let ftp_server = Arc::new(<Server as FTPRunner>::new(path.clone(), bind_ip.clone(), port));
+        let ftp_server_c = ftp_server.clone();
+
+        spawned_servers.push(ftp_server.clone());
+
+        spawned_runners.push(
             tokio::spawn(async move {
-            FTPServerRunner::runner(ftp_server.deref()).await
+                FTPRunner::runner(ftp_server.deref()).await
             })
         );
 
-        let port = cli_args.ftp.unwrap() as u16;
-
-        ftp_server_c.start(path.clone(), bind_ip.clone(), port);
+        spawned_runners.push(
+            tokio::spawn(async move {
+                let _ = ftp_server_c.start();
+            })
+        );
     }
 
 
@@ -153,21 +158,23 @@ async fn main() {
     // HTTP from here on
     // 
     // Spin the runners to wait for any potential server start
-    let http_server = Arc::new(<Server as HTTPServerRunner>::new());
-    let http_server_c = http_server.clone();
-
-    servers.push(http_server.clone());
-
     if cli_args.http.is_some() {
-        spawned_servers.push(
+        let port = cli_args.http.unwrap() as u16;
+        let server = Arc::new(<Server as HTTPRunner>::new(path.clone(), bind_ip.clone(), port));
+        let server_c = server.clone();
+
+        spawned_servers.push(server.clone());
+
+        spawned_runners.push(
             tokio::spawn(async move {
-            HTTPServerRunner::runner(http_server.deref()).await
+                HTTPRunner::runner(server).await
             })
         );
-
-        let port = cli_args.http.unwrap() as u16;
-
-        http_server_c.start(path.clone(), bind_ip.clone(), port);
+        spawned_runners.push(
+            tokio::spawn(async move {
+                let _ = server_c.start();
+            })
+        );
     }
 
 
@@ -176,37 +183,39 @@ async fn main() {
     // DHCP from here on
     // 
     // Spin the runners to wait for any potential server start
-    let dhcp_server = Arc::new(<Server as DHCPServerRunner>::new());
-    let dhcp_server_c = dhcp_server.clone();
-
-    servers.push(dhcp_server.clone());
-
     if cli_args.dhcp {
-        spawned_servers.push(
+        let dhcp_server = Arc::new(<Server as DHCPServerRunner>::new());
+        let dhcp_server_c = dhcp_server.clone();
+
+        spawned_servers.push(dhcp_server.clone());
+        spawned_runners.push(
             tokio::spawn(async move {
             DHCPServerRunner::runner(dhcp_server.deref()).await
             })
         );
 
-        dhcp_server_c.start(path.clone(), bind_ip.clone(), 0);
+        let _ = dhcp_server_c.start();
     }
+    //////////////////////////////////////////////////////////////////////
+    //////////////////////////////////////////////////////////////////////
 
-    //////////////////////////////////////////////////////////////////////
-    //////////////////////////////////////////////////////////////////////
+    if spawned_runners.iter().count() == 0 {
+        error!("No server(s) specified. Run with -h for more info...");
+        return;
+    }
 
     // Set up a handler for Ctrl+C signal
     ctrlc::set_handler(move || {
         // Handle Ctrl+C signal here
         warn!("Ctrl+C received. Closing connections and exiting.");
-        // You can perform cleanup operations here before exiting
-
-        for mut server in &mut servers {
+        // Perform cleanup operations here before exiting
+        for server in &mut spawned_servers {
             server.terminate();
         }
 
     }).expect("Error setting Ctrl+C handler");
     info!("Press Ctrl+C to exit.");
 
-    futures::future::join_all(spawned_servers).await;
+    futures::future::join_all(spawned_runners).await;
     return;
 }
